requirements.txt
fastapi==0.110
uvicorn==0.30
openai>=1.14
redis==5.0
pyyaml==6.0
